requirements.txt
anvil
anvil-app-server

nextcord
pyyaml
discord-py-slash-command

sqlalchemy
